+++ fs-storage/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod base_storage;
 pub mod file_storage;
+pub mod jni_file_storage;
 pub mod monoid;
 mod utils;
 pub const ARK_FOLDER: &str = ".ark";
